requirements.txt
pandas>=1.1.1
lasio>=0.29
streamlit
matplotlib.pyplot
